moved each component out of index.jsx in folder
