+++ src/components/PredictionSlider.jsx
